+++ pkg/router/template/plugin.go
@@ -3,7 +3,6 @@ package templaterouter
 import (
 	"crypto/md5"
 	"fmt"
-	"os"
 	"path/filepath"
 	"strconv"
 	"text/template"
@@ -91,30 +90,10 @@ type routerInterface interface {
 	Commit()
 }
 
-func env(name, defaultValue string) string {
-	if envValue := os.Getenv(name); envValue != "" {
-		return envValue
-	}
-
-	return defaultValue
-}
-
 // NewTemplatePlugin creates a new TemplatePlugin.
 func NewTemplatePlugin(cfg TemplatePluginConfig, lookupSvc ServiceLookup) (*TemplatePlugin, error) {
 	templateBaseName := filepath.Base(cfg.TemplatePath)
-	globalFuncs := template.FuncMap{
-		"endpointsForAlias":        endpointsForAlias,        //returns the list of valid endpoints
-		"processEndpointsForAlias": processEndpointsForAlias, //returns the list of valid endpoints after processing them
-		"env":          env,          //tries to get an environment variable if it can't return a default
-		"matchPattern": matchPattern, //anchors provided regular expression and evaluates against given string
-		"isInteger":    isInteger,    //determines if a given variable is an integer
-		"matchValues":  matchValues,  //compares a given string to a list of allowed strings
-
-		"genSubdomainWildcardRegexp": genSubdomainWildcardRegexp, //generates a regular expression matching the subdomain for hosts (and paths) with a wildcard policy
-		"generateRouteRegexp":        generateRouteRegexp,        //generates a regular expression matching the route hosts (and paths)
-		"genCertificateHostName":     genCertificateHostName,     //generates host name to use for serving/matching certificates
-	}
-	masterTemplate, err := template.New("config").Funcs(globalFuncs).ParseFiles(cfg.TemplatePath)
+	masterTemplate, err := template.New("config").Funcs(helperFunctions).ParseFiles(cfg.TemplatePath)
 	if err != nil {
 		return nil, err
 	}
